moved the enums to their own file & improvements
